Update ServoHack to toggle USB ports
On the Raspberry Pi 4, we can use uhubctl to toggle the state
of the USB Hubs. This effectively means that we can toggle the
5V power output of USB devices.

This is a much better approach than rebooting, as the robot will
remain on, broadcasting WiFi and can restart code almost immediately.

A bit of a general cleanup on the code too.